# Commit Message Fix build return values on buildProject function
This change updates the buildProject function to return two values. The first one is a slice of BuildError pointers that contain build errors, and the second one is a simple error in case of any internal problem during building.

The reason behind this change is to have a clearer understanding of the output of the build function, where it is possible to identify if any actual build error occurred, or if any internal error happened.