feat(psi): enable reading and writing to a file handle stored in an fs.FS
This commit adds a new feature to the `psi` package, allowing the ability to read and write from a file stored in an `fs.FS`. Specifically, the `SourceFile` struct has a new field `handle` of type `repofs.FileHandle`, and two new methods `Load` and `Replace`, which respectively read in the contents from the file handle and parse them into the `SourceFile` structure and write back out to the file handle from the contents of the `SourceFile` structure.

Additionally, the constructor for `SourceFile` has been updated to take in the fileset, name, and file handle.

A new file `handle.go` has been added to the `repofs` package, which contains an interface `FileHandle`. This interface abstracts reading and writing from the file handle to allow switching between using an `fs.FS` and a file system path.

Finally, a struct `FsFileHandle` has been added to the `repofs` package, which implements the `FileHandle` interface and allows reading and writing from a file stored in an `fs.FS`.