feat: add reindex command to root command
This commit adds the `reindex` command to the root command of the application. The `reindex` command invokes the `Reindex` method of the `Project` struct. The commit also adds the `utils` package to the project containing a `MarkdownTree` function to format JSON payloads as markdown. This function is used to format the output of the `SendToGPT` function in the `gpt` package. The formatting of the output is now performed by `utils.ParseMarkdown` instead of `ParseMarkdown` in the `gpt` package.