# Commit Message
## Implement initial design and structure for VTS package

- Remove unnecessary TODO comments in vts.go
- Sketch out the design of the VTS package
- Define the properties and behaviors of VTSQuery, VTSEdgeType, VTSNodeType, VTSFile, VTSPackage, VTSScope, VTSEdge, and VTSNode
- Add descriptive comments explaining the purpose and usage of the VTS package

The objective of these changes is to provide a solid foundation for implementing the Virtual Type System (VTS) in Go. The VTS package will serve as a generic type system that can represent any type of data structure. It will enable searching and traversing the VTS tree based on various criteria. By defining the properties and behaviors of VTSQuery, VTSEdgeType, VTSNodeType, VTSFile, VTSPackage, VTSScope, VTSEdge, and VTSNode, we establish the necessary components for building a flexible and comprehensive type system. The removal of unnecessary TODO comments ensures a clearer and more concise codebase. With this initial design and structure in place, we can proceed with further implementation and expansion of the VTS package.